Add assign
Again, implemented as a Python proc rather than a special case in eval.

You can write:
(assign 'a (+ 1 10))

instead of:
(define a (+ 1 10))

The only change to core is making 'token syntactic sugar for "token" if
token is an unquoted symbol.

It does depend on assign being able to receive env, obviously.